+++ src/api/pets.ts
@@ -14,5 +14,26 @@ const getPets = async () => {
   return response.data;
 };
 
-export { getPets };
-export type { PetsResponse };
+type petCreateRequest = Omit<Pet, "id" | "images"> & {
+  images: string[]; // image id
+};
+interface petCreateResponse extends Pet {}
+
+const token: string = "";
+const postPetCreate = async (
+  data: petCreateRequest
+): Promise<petCreateResponse> => {
+  const response = await axios.post<petCreateResponse>(
+    `${import.meta.env.VITE_API_URL}/pets/create`,
+    data,
+    {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    }
+  );
+  return response.data;
+};
+
+export { getPets, postPetCreate };
+export type { PetsResponse, petCreateRequest, petCreateResponse };
