+++ src/hooks/mutation/useCreatePet.ts
@@ -1,4 +1,4 @@
-import { petCreateRequest, postPetCreate } from "@/api/createPets";
+import { petCreateRequest, postPetCreate } from "@/api/pets";
 import { useMutation } from "@tanstack/react-query";
 
 const useCreatePet = () => {
